Move event APIs to top level file
