+++ gcal_sync/const.py
@@ -1,0 +1,7 @@
+"""Constants for gcal_sync."""
+
+# Local storage constants
+EVENT_SYNC = "event_sync"
+EVENTS = "events"
+TIMEZONE = "timezone"
+SYNC_TOKEN = "sync_token"
